Fixes after review part 3 Refactor RenderAmount to component Restructure the files
