+++ zarofeladat/src/utils/utils.js
@@ -1,17 +1,5 @@
-import { Typography } from '@mui/material';
-
 export const DISPLAY_INLINE_FLEX = { display: 'inline-flex' };
 export const DISPLAY_NONE = { display: 'none' };
-
-export function renderAmount(a) {
-  if (a < 0) {
-    return <Typography color="red">{a}</Typography>;
-  }
-  if (a > 0) {
-    return <Typography color="green">{a}</Typography>;
-  }
-  return <Typography>{a}</Typography>;
-}
 
 export function formatDate(dateString) {
   return new Date(dateString).toUTCString();
@@ -30,16 +18,6 @@ export function sortByDate(a, b) {
 export function basicValidator(value) {
   if (!value && value.length === 0) {
     return 'Field must not be empty';
-  }
-  return '';
-}
-
-export function numberValidator(value) {
-  if (!value && value.length === 0) {
-    return 'Field must not be empty';
-  }
-  if (Number.isNaN(Number(value))) {
-    return 'Value must be a number';
   }
   return '';
 }
